Code Re-Structured
