+++ lib/src/Screens/chat_screen.dart
@@ -1,17 +1,19 @@
 // ignore_for_file: prefer_const_constructors_in_immutables, use_key_in_widget_constructors, no_logic_in_create_state, non_constant_identifier_names
 
 import 'dart:io';
-import 'package:apptex_chat/src/Models/ChatModel.dart';
 import 'package:apptex_chat/src/Models/MessageModel.dart';
-import 'package:apptex_chat/src/Screens/full_screen_image.dart';
 import 'package:apptex_chat/src/Widgets/custom_animation.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import '../Controllers/chat_conrtroller.dart';
 import '../Controllers/contants.dart';
+import '../Widgets/image_bubble.dart';
+import '../Widgets/message_bubble.dart';
+import '../Widgets/profile.dart';
+import '../Widgets/replied_to.dart';
+import '../Widgets/swipe_to_reply.dart';
 
 class ChatScreen extends StatelessWidget {
   final String chatroomID;
@@ -447,211 +449,6 @@ class ChatScreen extends StatelessWidget {
   }
 }
 
-class RepliedToWidget extends StatelessWidget {
-  RepliedToWidget(
-      {Key? key,
-      required this.showCloseButton,
-      required this.chatController,
-      required this.messageId,
-      required this.myID,
-      required this.title,
-      this.messegedByMe = true})
-      : super(key: key);
-
-  final bool showCloseButton;
-  final ChatController chatController;
-  final String messageId;
-  final String myID;
-  final String title;
-  final bool messegedByMe;
-
-  @override
-  Widget build(BuildContext context) {
-    final size = MediaQuery.of(context).size;
-    print(messageId);
-
-    MessageModel model = chatController.messages.firstWhere(
-        (element) => element.uid == messageId,
-        orElse: () => MessageModel());
-    final bool isMine = model.sendBy == myID;
-
-    return Container(
-      width: size.width,
-      margin: showCloseButton
-          ? const EdgeInsets.only(top: 8.0)
-          : const EdgeInsets.all(4),
-      decoration: BoxDecoration(
-        color: showCloseButton ? Colors.white : Colors.black26.withOpacity(0.1),
-        borderRadius: showCloseButton
-            ? BorderRadius.circular(12)
-            : BorderRadius.circular(16),
-      ),
-      child: IntrinsicHeight(
-        child: Row(
-          children: [
-            Container(
-              width: 4,
-              margin: const EdgeInsets.only(top: 4.5, bottom: 4.5, left: 4),
-              decoration: BoxDecoration(
-                  color: showCloseButton || !messegedByMe
-                      ? kprimary1
-                      : Colors.white,
-                  borderRadius:
-                      const BorderRadius.horizontal(left: Radius.circular(16))),
-            ),
-            Expanded(
-              child: SingleChildScrollView(
-                physics: const NeverScrollableScrollPhysics(),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.only(
-                          left: 16.0, right: 8, top: 8, bottom: 4),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Text(
-                            isMine ? 'You' : title,
-                            style: TextStyle(
-                                color: showCloseButton || !messegedByMe
-                                    ? kprimary1
-                                    : Colors.white,
-                                fontWeight: FontWeight.w500,
-                                fontSize: 13),
-                          ),
-                          Visibility(
-                            visible: showCloseButton,
-                            child: GestureDetector(
-                              onTap: () {
-                                chatController.replyMessage.value = null;
-                              },
-                              child: CircleAvatar(
-                                backgroundColor: Colors.grey.shade300,
-                                maxRadius: 10,
-                                child: const Icon(
-                                  Icons.close_rounded,
-                                  size: 16,
-                                  color: Colors.black,
-                                ),
-                              ),
-                            ),
-                          )
-                        ],
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(
-                          left: 16.0, right: 16.0, bottom: 8.0),
-                      child: Text(
-                        model.message,
-                        maxLines: 2,
-                        overflow: TextOverflow.ellipsis,
-                        style: TextStyle(
-                            color: showCloseButton
-                                ? Colors.grey
-                                : messegedByMe
-                                    ? Colors.grey.shade200
-                                    : Colors.black54,
-                            fontSize: 12),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class MessageBubble extends StatelessWidget {
-  final bool isMine;
-  final MessageModel model;
-  final String profileUrl;
-  final DateTime msgDate;
-  final ChatController chatController;
-  final String myID;
-  final String title;
-
-  const MessageBubble(
-      {required this.isMine,
-      required this.model,
-      required this.profileUrl,
-      required this.msgDate,
-      required this.chatController,
-      Key? key,
-      required this.myID,
-      required this.title})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    double radius = 16;
-    return Row(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        mainAxisAlignment:
-            isMine ? MainAxisAlignment.end : MainAxisAlignment.start,
-        children: [
-          if (!isMine) Container(width: 3),
-          if (isMine) getdate(msgDate),
-          if (!isMine) ProfilePic(profileUrl, size: 26),
-          IntrinsicWidth(
-              child: Container(
-            constraints: BoxConstraints(
-                minWidth: 50,
-                maxWidth: MediaQuery.of(context).size.width * 0.6),
-            margin: const EdgeInsets.symmetric(vertical: 4, horizontal: 10),
-            // padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-            decoration: BoxDecoration(
-              color: isMine ? kprimary1 : kprimary2,
-              borderRadius: BorderRadius.only(
-                topLeft:
-                    isMine ? Radius.circular(radius) : Radius.circular(radius),
-                bottomLeft:
-                    isMine ? Radius.circular(radius) : const Radius.circular(0),
-                bottomRight:
-                    isMine ? const Radius.circular(0) : Radius.circular(radius),
-                topRight:
-                    isMine ? Radius.circular(radius) : Radius.circular(radius),
-              ),
-            ),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                model.repliedTo != null
-                    ? RepliedToWidget(
-                        myID: myID,
-                        title: title,
-                        messegedByMe: isMine,
-                        messageId: model.repliedTo!,
-                        showCloseButton: false,
-                        chatController: chatController)
-                    : const SizedBox(),
-                Padding(
-                  padding:
-                      const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                  child: Text(
-                    model.message,
-                    style: TextStyle(
-                        color: isMine ? kWhite : Colors.grey.shade800,
-                        fontWeight: FontWeight.w500,
-                        fontSize: 13),
-                  ),
-                ),
-              ],
-            ),
-          )),
-          //  if (isMine) ProfilePic(profileUrl, size: 30),
-          //if (isMine) Container(width: 3),
-          if (!isMine) getdate(msgDate),
-        ]);
-  }
-}
 
 getdate(var msgDate) {
   String date = getChatDate(msgDate);
@@ -663,340 +460,4 @@ getdate(var msgDate) {
       style: myStyle(9, false, color: Colors.grey.shade600),
     ),
   );
-}
-
-class ImageBubble extends StatelessWidget {
-  final bool isMine;
-  final String message;
-  final String profileUrl;
-  final DateTime msgDate;
-  const ImageBubble(this.isMine, this.message, this.profileUrl, this.msgDate,
-      {Key? key})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    double boxSize = 180;
-    return Row(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        mainAxisAlignment:
-            isMine ? MainAxisAlignment.end : MainAxisAlignment.start,
-        children: [
-          if (!isMine) Container(width: 3),
-          if (!isMine) ProfilePic(profileUrl, size: 30),
-          if (isMine) getdate(msgDate),
-          IntrinsicWidth(
-              child: Hero(
-            tag: message,
-            child: GestureDetector(
-              onTap: () {
-                Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => FullScreenImage(
-                              imgUrl: message,
-                            )));
-              },
-              child: Container(
-                height: boxSize,
-                width: boxSize,
-                constraints: BoxConstraints(
-                    minWidth: 50,
-                    maxWidth: MediaQuery.of(context).size.width / 1.3),
-                margin: const EdgeInsets.symmetric(vertical: 4, horizontal: 10),
-                padding:
-                    const EdgeInsets.symmetric(vertical: 9, horizontal: 10),
-                decoration: BoxDecoration(
-                  image: DecorationImage(
-                      image: NetworkImage(message), fit: BoxFit.cover),
-                  color: isMine ? kprimary1 : kprimary2,
-                  borderRadius: BorderRadius.only(
-                    topLeft: isMine
-                        ? const Radius.circular(20)
-                        : const Radius.circular(0),
-                    bottomLeft: isMine
-                        ? const Radius.circular(20)
-                        : const Radius.circular(10),
-                    bottomRight: isMine
-                        ? const Radius.circular(10)
-                        : const Radius.circular(20),
-                    topRight: isMine
-                        ? const Radius.circular(0)
-                        : const Radius.circular(20),
-                  ),
-                ),
-              ),
-            ),
-          )),
-          if (!isMine) getdate(msgDate),
-        ]);
-  }
-}
-
-// ignore: must_be_immutable
-class ProfilePic extends StatelessWidget {
-  ProfilePic(this.url, {this.size = 48, this.radius = 100, Key? key})
-      : super(key: key);
-  String url;
-
-  final double size;
-  final double radius;
-  @override
-  Widget build(BuildContext context) {
-    url = url.length <= 7 ? "" : url;
-    bool isNotUrl = url.length <= 7;
-
-    double imsix = size;
-    return ClipRRect(
-      borderRadius: BorderRadius.circular(radius),
-      child: isNotUrl
-          ? Container(
-              color: kprimary1,
-              width: imsix,
-              height: imsix,
-              child: const Icon(
-                Icons.person,
-                size: 20,
-              ))
-          : Container(
-              width: imsix,
-              height: imsix,
-              color: Colors.grey.shade300,
-              child: Image(
-                image: NetworkImage(
-                  url,
-                ),
-                alignment: Alignment.center,
-                height: double.infinity,
-                width: double.infinity,
-                fit: BoxFit.cover,
-              ),
-            ),
-    );
-  }
-}
-
-class SwipeToReply extends StatefulWidget {
-  /// Child widget for which you want to have horizontal swipe action
-  /// @required parameter
-  final Widget child;
-
-  /// Duration value to define animation duration
-  /// if not passed default Duration(milliseconds: 150) will be taken
-  final Duration animationDuration;
-
-  /// Icon that will be displayed beneath child widget when swipe right
-  final IconData iconOnRightSwipe;
-
-  /// Widget that will be displayed beneath child widget when swipe right
-  final Widget? rightSwipeWidget;
-
-  /// Icon that will be displayed beneath child widget when swipe left
-  final IconData iconOnLeftSwipe;
-
-  /// Widget that will be displayed beneath child widget when swipe right
-  final Widget? leftSwipeWidget;
-
-  /// double value defining size of displayed icon beneath child widget
-  /// if not specified default size 26 will be taken
-  final double iconSize;
-
-  /// color value defining color of displayed icon beneath child widget
-  ///if not specified primaryColor from theme will be taken
-  final Color? iconColor;
-
-  /// Double value till which position child widget will get animate when swipe left
-  /// or swipe right
-  /// if not specified 0.3 default will be taken for Right Swipe &
-  /// it's negative -0.3 will bve taken for Left Swipe
-  final double offsetDx;
-
-  /// bool value till which indicate whether the it is swip right or left
-  /// true for right swipe while false for left
-  final bool isRightSwipe;
-
-  /// callback which will be initiated at the end of child widget animation
-  /// when swiped right
-  /// if not passed swipe to right will be not available
-  final GestureDragEndCallback? onRightSwipe;
-
-  /// callback which will be initiated at the end of child widget animation
-  /// when swiped left
-  /// if not passed swipe to left will be not available
-  final GestureDragEndCallback? onLeftSwipe;
-
-  /// callback which will be initiated when user long press or hold
-  final VoidCallback? onHold;
-
-  const SwipeToReply({
-    Key? key,
-    required this.child,
-    required this.isRightSwipe,
-    this.onRightSwipe,
-    this.onLeftSwipe,
-    this.onHold,
-    this.iconOnRightSwipe = CupertinoIcons.arrowshape_turn_up_left_fill,
-    this.rightSwipeWidget,
-    this.iconOnLeftSwipe = CupertinoIcons.arrowshape_turn_up_right_fill,
-    this.leftSwipeWidget,
-    this.iconSize = 22.0,
-    this.iconColor = Colors.grey,
-    this.animationDuration = const Duration(milliseconds: 130),
-    this.offsetDx = 0.2,
-  }) : super(key: key);
-
-  @override
-  _SwipeToReplyState createState() => _SwipeToReplyState();
-}
-
-class _SwipeToReplyState extends State<SwipeToReply>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<Offset> _animation;
-  late Animation<double> _leftIconAnimation;
-  late Animation<double> _rightIconAnimation;
-  late GestureDragEndCallback _onSwipeLeft;
-  late GestureDragEndCallback _onSwipeRight;
-
-  @override
-  initState() {
-    super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: widget.animationDuration,
-    );
-    _animation = Tween<Offset>(
-      begin: const Offset(0.0, 0.0),
-      end: const Offset(0.0, 0.0),
-    ).animate(
-      CurvedAnimation(curve: Curves.decelerate, parent: _controller),
-    );
-    _leftIconAnimation = _controller.drive(
-      Tween<double>(begin: 0.0, end: 0.0),
-    );
-    _rightIconAnimation = _controller.drive(
-      Tween<double>(begin: 0.0, end: 0.0),
-    );
-    _onSwipeLeft = widget.onLeftSwipe ??
-        (details) {
-          throw ("Left Swipe Not Provided");
-        };
-
-    _onSwipeRight = widget.onRightSwipe ??
-        (details) {
-          throw ("Left Swipe Not Provided");
-        };
-    _controller.addListener(() {
-      setState(() {});
-    });
-  }
-
-  @override
-  dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  ///Run animation for child widget
-  ///[onRight] value defines animation Offset direction
-  void _runAnimation() {
-    //set child animation
-
-    _animation = Tween(
-      begin: const Offset(0.0, 0.0),
-      end:
-          Offset(widget.isRightSwipe ? widget.offsetDx : -widget.offsetDx, 0.0),
-    ).animate(
-      CurvedAnimation(curve: Curves.decelerate, parent: _controller),
-    );
-    //set back left/right icon animation
-    if (widget.isRightSwipe) {
-      _leftIconAnimation = Tween(begin: 0.0, end: 1.0).animate(
-        CurvedAnimation(curve: Curves.decelerate, parent: _controller),
-      );
-      _rightIconAnimation = Tween(begin: 0.0, end: 0.0).animate(
-        CurvedAnimation(curve: Curves.decelerate, parent: _controller),
-      );
-    } else {
-      _rightIconAnimation = Tween(begin: 0.0, end: 1.0).animate(
-        CurvedAnimation(curve: Curves.decelerate, parent: _controller),
-      );
-      _leftIconAnimation = Tween(begin: 0.0, end: 0.0).animate(
-        CurvedAnimation(curve: Curves.decelerate, parent: _controller),
-      );
-    }
-    //Forward animation
-    _controller.forward();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onLongPress: widget.onHold,
-      onHorizontalDragEnd: ((details) {
-        _controller.reverse().whenComplete(() {
-          if (widget.isRightSwipe) {
-            //keep left icon visibility to 0.0 until onRightSwipe triggers again
-            // _leftIconAnimation = _controller.drive(Tween(begin: 0.0, end: 0.0));
-            _onSwipeRight(details);
-          } else {
-            //keep right icon visibility to 0.0 until onLeftSwipe triggers again
-            // _rightIconAnimation = _controller.drive(Tween(begin: 0.0, end: 0.0));
-
-            _onSwipeLeft(details);
-          }
-        });
-      }),
-      onHorizontalDragStart: (details) {
-        if (widget.isRightSwipe) {
-          if (details.localPosition.dx > 2) {
-            _runAnimation();
-          }
-        } else {
-          if (details.localPosition.dx > -2) {
-            _runAnimation();
-          }
-        }
-      },
-      child: Stack(
-        alignment: Alignment.center,
-        fit: StackFit.passthrough,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              AnimatedOpacity(
-                opacity: _leftIconAnimation.value,
-                duration: widget.animationDuration,
-                curve: Curves.decelerate,
-                child: widget.rightSwipeWidget ??
-                    Icon(
-                      widget.iconOnRightSwipe,
-                      size: widget.iconSize,
-                      color:
-                          widget.iconColor ?? Theme.of(context).iconTheme.color,
-                    ),
-              ),
-              AnimatedOpacity(
-                opacity: _rightIconAnimation.value,
-                duration: widget.animationDuration,
-                curve: Curves.decelerate,
-                child: widget.leftSwipeWidget ??
-                    Icon(
-                      widget.iconOnLeftSwipe,
-                      size: widget.iconSize,
-                      color:
-                          widget.iconColor ?? Theme.of(context).iconTheme.color,
-                    ),
-              ),
-            ],
-          ),
-          SlideTransition(
-            position: _animation,
-            child: widget.child,
-          ),
-        ],
-      ),
-    );
-  }
 }
